move createEVM/create2EVM tests back to system.test.ts
