+++ core/tests/ts-integration/tests/evm.test.ts
@@ -106,98 +106,3 @@ describeEvm('EVM contract checks', () => {
         await expect(counterFactory.deploy(true)).rejects.toThrow('requested revert');
     });
 });
-
-describeEvm('createEVM/create2EVM functions checks', () => {
-    let testMaster: TestMaster;
-    let alice: zksync.Wallet;
-    let isETHBasedChain: boolean;
-    let baseTokenAddress: string;
-    let expectedL2Costs: bigint;
-
-    beforeAll(async () => {
-        testMaster = TestMaster.getInstance(__filename);
-        alice = testMaster.mainAccount();
-        baseTokenAddress = await alice._providerL2().getBaseTokenContractAddress();
-        isETHBasedChain = baseTokenAddress == zksync.utils.ETH_ADDRESS_IN_CONTRACTS;
-        const gasPrice = await scaledGasPrice(alice);
-        if (!isETHBasedChain) {
-            expectedL2Costs =
-                ((await alice.getBaseCost({
-                    gasLimit: maxL2GasLimitForPriorityTxs(testMaster.environment().priorityTxMaxGasLimit),
-                    gasPerPubdataByte: REQUIRED_L1_TO_L2_GAS_PER_PUBDATA_LIMIT,
-                    gasPrice
-                })) *
-                    140n) /
-                100n;
-        }
-    });
-
-    test('Should check that createEVM is callable and executes successfully', async () => {
-        if (!isETHBasedChain) {
-            const baseTokenDetails = testMaster.environment().baseToken;
-            const baseTokenMaxAmount = await alice.getBalanceL1(baseTokenDetails.l1Address);
-            await (await alice.approveERC20(baseTokenDetails.l1Address, baseTokenMaxAmount)).wait();
-        }
-
-        const contractInitcode: BytesLike = ethers.hexlify('0x69602a60005260206000f3600052600a6016f3');
-
-        const abi = ['function createEVM(bytes _initCode)'];
-
-        const iface = new ethers.Interface(abi);
-
-        const ContractDeployerCalldata: BytesLike = iface.encodeFunctionData('createEVM', [contractInitcode]);
-
-        const gasPrice = await scaledGasPrice(alice);
-        const l2GasLimit = maxL2GasLimitForPriorityTxs(testMaster.environment().priorityTxMaxGasLimit);
-
-        let priorityOpHandle = await alice.requestExecute({
-            contractAddress: await zksync.utils.CONTRACT_DEPLOYER_ADDRESS,
-            calldata: ContractDeployerCalldata,
-            l2GasLimit: l2GasLimit,
-            mintValue: isETHBasedChain ? 0n : expectedL2Costs,
-            overrides: {
-                gasPrice
-            }
-        });
-
-        await priorityOpHandle.waitL1Commit();
-
-        expect(priorityOpHandle).toBeAccepted();
-    });
-
-    test('Should check that create2EVM is callable and executes successfully', async () => {
-        if (!isETHBasedChain) {
-            const baseTokenDetails = testMaster.environment().baseToken;
-            const baseTokenMaxAmount = await alice.getBalanceL1(baseTokenDetails.l1Address);
-            await (await alice.approveERC20(baseTokenDetails.l1Address, baseTokenMaxAmount)).wait();
-        }
-
-        const contractInitcode: BytesLike = '0x69602a60005260206000f3600052600a6016f3';
-
-        const abi = ['function create2EVM(bytes32 _salt, bytes _initCode)'];
-
-        const iface = new ethers.Interface(abi);
-
-        const ContractDeployerCalldata: BytesLike = iface.encodeFunctionData('create2EVM', [
-            ethers.ZeroHash,
-            contractInitcode
-        ]);
-
-        const gasPrice = await scaledGasPrice(alice);
-        const l2GasLimit = maxL2GasLimitForPriorityTxs(testMaster.environment().priorityTxMaxGasLimit);
-
-        let priorityOpHandle = await alice.requestExecute({
-            contractAddress: await zksync.utils.CONTRACT_DEPLOYER_ADDRESS,
-            calldata: ContractDeployerCalldata,
-            l2GasLimit: l2GasLimit,
-            mintValue: isETHBasedChain ? 0n : expectedL2Costs,
-            overrides: {
-                gasPrice
-            }
-        });
-
-        await priorityOpHandle.waitL1Commit();
-
-        expect(priorityOpHandle).toBeAccepted();
-    });
-});
